+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v2.2.14
+
+* _TBA_
+
 # v2.2.13
 
 * useless `print()`s removed in metrics